+++ src/__tests__/lib/lolex.js
